product refactor and proptype added
